Common perf tests
