+++ .circleci/config.yml
@@ -60,7 +60,12 @@ jobs:
                   at: .
 
             - run:
-                  command: time node --expose-gc test/perf/index.js
+                  command: time node --expose-gc test/perf/index.js v4
+                  environment:
+                      PERSIST: true
+
+            - run:
+                  command: time node --expose-gc test/perf/index.js v5
                   environment:
                       PERSIST: true
