+++ .gitignore
@@ -7,7 +7,6 @@ notes.md
 lib
 test/babel-tests.js
 test/typescript/typescript-tests.js
-test/perf/perf.txt
 dist/
 .build*/
 .idea
